refactor(core): make core the single formatMemoryUsage, fixing the other two copies
The rollover fix landed in the CLI copy only, leaving the identical bug in
packages/core/src/utils/formatters.ts (shell truncation notices, memory
diagnostics) and packages/cli/src/serve/env-snapshot.ts (the daemon's
/workspace/env). The same byte count would have printed "1.0 MB" or
"1024.0 KB" depending on which path formatted it.

Core already re-exports formatters from its public index, so rather than
carry a third copy of the fix the other two now import it. Net -34 lines.
The case table moves to core alongside the implementation; the CLI test
keeps a smaller set covering the re-export, rollover included.

Verified by restoring main's core implementation: the two rollover cases
fail. The unrelated AnsiOutput "getFrameController is not a function"
failure reproduces on unmodified main.
